Improving the data preparation code by reducing the scope of the dataset.
This has been done introducing and cleaning a different dataset containing additional useful data about the country currency.
This new dataset has then been merged into the original one, so that we had a view of all the countries and the relative currencies.
From there, it was possible to filter out all non-Euro countries, which allowed us to prepare the data relevantly to the target audience and problem statement.

The cleaned dataframe as been saved as a new prepared .csv file.